requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
PyMySQL==1.1.0
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
twilio==8.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0
